requirements.txt
fastapi
uvicorn
anthropic
sentence-transformers
chromadb
google-genai
python-multipart
gitpython
python-dotenv
requests